debugged cpu.c + general updates

+++ dram.c
@@ -1,4 +1,5 @@
 #include "dram.h"
+#include "utils.h"
 
 DRAM dram;
 
@@ -12,7 +13,7 @@ Result load(u64 addr, u64 size){
     int index = addr - DRAM_BASE;
     u64 val = 0;
     for(int i = 0;i < (size/8);i++){
-        val |= (u64)dram.mem[index+i] << (i*8);
+        val |= (((u64)dram.mem[index+i]) << (i*8));
     }
     ret.value = val;
     return ret;
@@ -27,7 +28,7 @@ Result store(u64 addr, u64 size, u64 val){
     }
     int index = addr - DRAM_BASE;
     for(int i = 0;i < (size/8);i++){
-        dram.mem[index+i] = (val >> (i*8)) & 0xff;
+        dram.mem[index+i] = (u8) ((val >> (i*8)) & 0xff);
     }
     return ret;
 }
